Fix detection of Ruby.framework on Mac OS X 10.9
Mavericks has moved the ruby headers into Xcode.app, so checking for
them at the physical location will fail. Because of this we need to
remove checks that rely on the exact locations returned by RbConfig
(which mkmf and test -d will) and instead trust the compiler to append
the proper framework directory inside Xcode.app.

This is backwards compatible with at least 10.6 and 10.7.

It does not yet match the compiled Ruby.framework version to the
ruby-command given to configure, because -framework Ruby will always use
the latest version. So currently it will compile against 2.0 on 10.9 and
1.8 on pre-10.9 OS X.